ci: Refactor CI triggers and revert pull_request_target
To avoid unnecessary secrets exposure, we revert to the `pull_request`
event to trigger our CI. We keep `pull_request_target` only for adding
git-trailers.

Additionally, we now trigger the CI on PRs when they are opened or
synchronized, and not on each labeling event, preventing unwanted
triggers and skipped checks. We just keep labeled, to trigger the
tests if the PR is opened and labeled afterwards.

We also add an extra label `takeover` when there's an external PR (from
a forked instance), and we do that, because we would like to add the
trailers. This ends-up issuing another PR (silently) from a local
branch.

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>